requirements.txt
# LLMOps Workshop - Python Dependencies
# ======================================
# Install with: pip install -r requirements.txt

# Azure Identity (for RBAC authentication)
azure-identity>=1.15.0

# Azure AI Services
azure-ai-evaluation>=1.0.0b1
azure-ai-projects>=1.0.0b1
azure-search-documents>=11.4.0
azure-mgmt-cognitiveservices>=13.5.0

# Azure OpenAI
openai>=1.12.0

# Prompt Flow
promptflow>=1.10.0
promptflow-tools>=1.4.0
promptflow-azure>=1.10.0

# Web Frontend
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dotenv>=1.0.0
rich>=13.7.0
